Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/fetch/sec-metadata/sharedworker.tentative.https.sub-expected.txt

CONSOLE MESSAGE: line 13: ReferenceError: Can't find variable: SharedWorker

FAIL Untitled ReferenceError: Can't find variable: SharedWorker
